// console/src/generated/LoginParamDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LoginModeDto } from "./LoginModeDto";
import type { LoginParamKindDto } from "./LoginParamKindDto";

export type LoginParamDto = { name: string, kind: LoginParamKindDto, required: boolean, default_value: string | null, options: Array<string>, modes: Array<LoginModeDto>, };